feat: outfits system
Stores JSON strings into server kvp.
Outfit names are stored as "slot": name in kvp and loaded when the player is.
Appearance data is stored as "slot": appearance and loads when requested.

Once data has been loaded it is cached on the client and will not be requested unless cleared. Requires manual setup for non-ESX frameworks when using multicharacter systems.

No UI for outfits (yet).

Includes a bunch of code tweaks for organisation / simplicity. Old 'main' renamed to 'utils', and 'shops' is now 'main'.